refactor: move code out of lib.rs
There's a lot of unrelated stuff in there, and in combination with
re-exports, global lint configs, module declarations, and other stuff,
it's just easier to move the actual code into separate files.
