+++ cot-codegen/src/lib.rs
@@ -1,5 +1,3 @@
-extern crate self as cot_codegen;
-
 pub mod expr;
 pub mod model;
 #[cfg(feature = "symbol-resolver")]
